pyproject.toml
[tool.poetry]
name = "quantlib-server"
version = "0.1.0"
description = "A server setup with the QuantLib Python module available"
authors = ["eisvogel.eth <eisvogel.eth>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.11"
quantlib = "*"
matplotlib = "*"
numpy = "*"
scipy = "*"
pandas = "*"
ipykernel = "^6.29.5"
jupyter = "^1.1.1"
nbformat = "^5.10.4"
psycopg2-binary = "^2.9.10"
notebook = "^7.3.2"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
